+++ packages/components/legend/rollup.config.js
@@ -4,9 +4,9 @@ import path from 'path';
 import {terser} from 'rollup-plugin-terser';
 import analyze from 'rollup-plugin-analyzer';
 import cleanup from 'rollup-plugin-cleanup';
-import commonjs from 'rollup-plugin-commonjs';
+import commonjs from '@rollup/plugin-commonjs';
 import json from '@rollup/plugin-json';
-import resolve from 'rollup-plugin-node-resolve';
+import resolve from '@rollup/plugin-node-resolve';
 import svelte from 'rollup-plugin-svelte';
 
 import * as _ from 'lamb';
@@ -57,7 +57,9 @@ const makeBrowserConfig = _.pipe([
 		plugins: [
 			resolve(),
 			commonjs(),
-			svelte(),
+			svelte({
+				emitCss: false
+			}),
 			json(),
 			removeComments,
 			// buble({
@@ -100,7 +102,9 @@ const cjsConfig = {
 	plugins: [
 		resolve(),
 		commonjs(),
-		svelte(),
+		svelte({
+			emitCss: false
+		}),
 		json(),
 		removeComments,
 	],