+++ packages/components/time_region_value/rollup.config.js
@@ -4,8 +4,9 @@ import path from 'path';
 import {terser} from 'rollup-plugin-terser';
 import analyze from 'rollup-plugin-analyzer';
 import cleanup from 'rollup-plugin-cleanup';
-import commonjs from 'rollup-plugin-commonjs';
-import resolve from 'rollup-plugin-node-resolve';
+import commonjs from '@rollup/plugin-commonjs';
+import json from '@rollup/plugin-json';
+import resolve from '@rollup/plugin-node-resolve';
 import svelte from 'rollup-plugin-svelte';
 
 import * as _ from 'lamb';
@@ -54,9 +55,11 @@ const makeBrowserConfig = _.pipe([
 		plugins: [
 			resolve(),
 			commonjs(),
-			svelte(),
+			json(),
+			svelte({
+				emitCss: false
+			}),
 			removeComments,
-			// json(),
 			// buble({
 			//	 transforms: { dangerousForOf: true }
 			// }),
@@ -97,7 +100,10 @@ const cjsConfig = {
 	plugins: [
 		resolve(),
 		commonjs(),
-		svelte(),
+		json(),
+		svelte({
+			emitCss: false
+		}),
 		removeComments,
 	],
 	treeshake