+++ packages/docs/site/rollup.config.js
@@ -28,8 +28,10 @@ export default {
 				'process.env.NODE_ENV': JSON.stringify(mode)
 			}),
 			svelte({
-				dev,
-				hydratable: true,
+				compilerOptions: {
+					dev,
+					hydratable: true,
+				},
 				emitCss: true,
 			}),
 			resolve({
@@ -89,8 +91,10 @@ export default {
 				'process.env.NODE_ENV': JSON.stringify(mode)
 			}),
 			svelte({
-				generate: 'ssr',
-				dev,
+				compilerOptions: {
+					dev,
+					generate: 'ssr',
+				},
 			}),
 			resolve({
 				dedupe: ['svelte']