+++ packages/tools/dev/rollup.config.js
@@ -1,11 +1,12 @@
-import path from "path";
-import analyze from "rollup-plugin-analyzer";
-import cleanup from "rollup-plugin-cleanup";
-import commonjs from "rollup-plugin-commonjs";
-import resolve from "rollup-plugin-node-resolve";
+import path from 'path';
 
-import {makeBanner} from "./src/common";
-import pkg from "./package.json";
+import analyze from 'rollup-plugin-analyzer';
+import cleanup from 'rollup-plugin-cleanup';
+import commonjs from '@rollup/plugin-commonjs';
+import resolve from '@rollup/plugin-node-resolve';
+
+import {makeBanner} from './src/common';
+import pkg from './package.json';
 
 const analyzer = analyze({
 	limit: 15,
@@ -30,7 +31,7 @@ const cjsConfig = {
 	output: {
 		banner,
 		file: pkg.main,
-		format: "cjs",
+		format: 'cjs',
 		indent: false
 	},
 	plugins: [