+++ packages/tools/file/rollup.config.js
@@ -1,12 +1,13 @@
-import path from "path";
-import analyze from "rollup-plugin-analyzer";
-import cleanup from "rollup-plugin-cleanup";
-import commonjs from "rollup-plugin-commonjs";
-import resolve from "rollup-plugin-node-resolve";
+import path from 'path';
 
-import {makeBanner} from "@svizzle/dev";
+import analyze from 'rollup-plugin-analyzer';
+import cleanup from 'rollup-plugin-cleanup';
+import commonjs from '@rollup/plugin-commonjs';
+import resolve from '@rollup/plugin-node-resolve';
 
-import pkg from "./package.json";
+import {makeBanner} from '@svizzle/dev';
+
+import pkg from './package.json';
 
 const analyzer = analyze({
 	limit: 15,
@@ -31,7 +32,7 @@ const cjsConfig = {
 	output: {
 		banner,
 		file: pkg.main,
-		format: "cjs",
+		format: 'cjs',
 		indent: false
 	},
 	plugins: [