+++ packages/components/choropleth/CHANGELOG.md
@@ -1,3 +1,11 @@
+## `@svizzle/choropleth` v0.6.3 (next)
+
+- updated some dependencies:
+	- `d3-geo`
+	- `@rollup/plugin-commonjs`
+	- `@rollup/plugin-node-resolve`
+	- `rollup-plugin-svelte` (for now using `emitCss: false`)
+
 ## `@svizzle/choropleth` v0.6.2
 
 - show the `message` only if `projection` is `undefined` OR the internal geojson has an empty `features` property