+++ packages/tools/dom/CHANGELOG.md
@@ -1,3 +1,7 @@
+## `@svizzle/dom` v0.4.3 (next)
+
+- updated some dev dependencies
+
 ## `@svizzle/dom` v0.4.2
 
 - depends on a newer `@svizzle/utils` because of `prepend` renamed to `makePrefixed`