+++ packages/tools/file/CHANGELOG.md
@@ -1,3 +1,7 @@
+## `@svizzle/file` v0.9.1 (next)
+
+- updated some dev dependencies
+
 ## `@svizzle/file` v0.9.0
 
 - add `readDirFiles`