+++ packages/tools/geo/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## `@svizzle/geo` v0.6.2 (next)
 
-- Updated `@turf/*` to version `6.3.0`
+- updated `@turf/*` to version `6.3.0`
+- updated some dev dependencies
 
 ## `@svizzle/geo` v0.6.1