+++ packages/tools/geometry/CHANGELOG.md
@@ -1,3 +1,7 @@
+## `@svizzle/geometry` v0.3.2 (next)
+
+- updated some dev dependencies
+
 ## `@svizzle/geometry` v0.3.1
 
 - build: remove comments from `.mjs` files